requirements.txt
appdirs==1.4.3
asgiref==3.2.6
Babel==2.8.0
cfgv==3.1.0
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
Django==2.2.11
django-appconf==1.0.3
django-auditlog==0.4.7
django-daterange-filter==1.3.0
django-dotenv==1.4.2
django-extensions==2.2.8
django-filter==2.2.0
django-fsm==2.7.0
django-fsm-admin==1.2.4
django-fsm-log==1.6.2
django-import-export==2.0.2
django-jsonfield==1.4.0
django-phonenumber-field==4.0.0
django-taggit==1.2.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
filelock==3.0.12
graphviz==0.13.2
identify==1.4.12
importlib-metadata==1.5.0
jdcal==1.4.1
Markdown==3.2.1
MarkupPy==1.14
nodeenv==1.3.5
odfpy==1.4.1
openpyxl==3.0.3
phonenumbers==8.12.0
pre-commit==2.2.0
python-dateutil==2.6.0
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
toml==0.10.0
virtualenv==20.0.13
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
